+++ packages/styleguide/stories/Foundations/Layouts/Grid/Grid.stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { select } from '@storybook/addon-knobs';
-
 import { LayoutGrid, Column } from '@codecademy/gamut/src';
 
 import {
@@ -10,33 +9,32 @@ import {
   StoryTemplate,
   StoryDescription,
 } from '../../../Templating';
-
-import { Box, Container } from '../Elements';
+import { Content, Container } from './Elements';
 
 export default decoratedStories('Foundations', 'Layouts', 'Grid', LayoutGrid);
 
 const KitchenSinkColumns = () => (
   <>
     <Column size={12}>
-      <Box>12</Box>
+      <Content>12</Content>
     </Column>
     <Column size={6}>
-      <Box>6</Box>
+      <Content>6</Content>
     </Column>
     <Column size={6}>
-      <Box>6</Box>
+      <Content>6</Content>
     </Column>
     <Column size={3}>
-      <Box>3</Box>
+      <Content>3</Content>
     </Column>
     <Column size={3}>
-      <Box>3</Box>
+      <Content>3</Content>
     </Column>
     <Column size={3}>
-      <Box>3</Box>
+      <Content>3</Content>
     </Column>
     <Column size={3}>
-      <Box>3</Box>
+      <Content>3</Content>
     </Column>
   </>
 );
@@ -55,81 +53,84 @@ export const basics = decoratedStory(() => {
         ones on large viewports.
       </StoryDescription>
       <Container>
-        <LayoutGrid columnGap="sm" rowGap="sm">
+        <LayoutGrid
+          columnGap={{ xs: 'sm', lg: 'lg' }}
+          rowGap={{ xs: 'sm', lg: 'lg' }}
+        >
           <Column size={12}>
-            <Box>12</Box>
+            <Content>12</Content>
           </Column>
           <Column size={6}>
-            <Box>6</Box>
+            <Content>6</Content>
           </Column>
           <Column size={6}>
-            <Box>6</Box>
+            <Content>6</Content>
           </Column>
           <Column size={3}>
-            <Box>3</Box>
+            <Content>3</Content>
           </Column>
           <Column size={3}>
-            <Box>3</Box>
+            <Content>3</Content>
           </Column>
           <Column size={3}>
-            <Box>3</Box>
+            <Content>3</Content>
           </Column>
           <Column size={3}>
-            <Box>3</Box>
+            <Content>3</Content>
           </Column>
           <Column size={2}>
-            <Box>2</Box>
+            <Content>2</Content>
           </Column>
           <Column size={2}>
-            <Box>2</Box>
+            <Content>2</Content>
           </Column>
           <Column size={2}>
-            <Box>2</Box>
+            <Content>2</Content>
           </Column>
           <Column size={2}>
-            <Box>2</Box>
+            <Content>2</Content>
           </Column>
           <Column size={2}>
-            <Box>2</Box>
+            <Content>2</Content>
           </Column>
           <Column size={2}>
-            <Box>2</Box>
+            <Content>2</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
           <Column size={1}>
-            <Box>1</Box>
+            <Content>1</Content>
           </Column>
         </LayoutGrid>
       </Container>
@@ -138,21 +139,22 @@ export const basics = decoratedStory(() => {
 });
 
 export const editable = decoratedStory(() => {
-  const rowGap = select('Row Gap', [undefined, 'sm', 'md', 'lg', 'xl'], 'sm');
   const columnGap = select(
     'Column Gap',
     [undefined, 'sm', 'md', 'lg', 'xl'],
     'sm'
   );
+  const rowGap = select('Row Gap', [undefined, 'sm', 'md', 'lg', 'xl'], 'sm');
 
   return (
     <StoryTemplate status={StoryStatus.Ready}>
       <StoryDescription>
-        Columns may be offset from the left by numbers within the [0-12] scale.
-        As with column and row gaps, those numbers may be static or responsive.
+        You can specify both the <em>column</em> and <em>row</em> gap amounts
+        manually to override the default responsive layout. They both default to
+        0 if not specified.
       </StoryDescription>
       <Container>
-        <LayoutGrid rowGap={rowGap} columnGap={columnGap}>
+        <LayoutGrid {...{ columnGap, rowGap }}>
           <KitchenSinkColumns />
         </LayoutGrid>
       </Container>
